a few fixes to integrate better with k3d recreation

+++ docs/assets/scripts/developer/k3d-dev.sh
@@ -4,6 +4,10 @@ function run() {
   ssh -i ~/.ssh/${KeyName}.pem -o StrictHostKeyChecking=no -o IdentitiesOnly=yes ubuntu@${PublicIP} "$@"
 }
 
+function getPrivateIP2() {
+  echo `aws ec2 describe-instances --output json --no-cli-pager --instance-ids ${InstId} | jq -r '.Reservations[0].Instances[0].NetworkInterfaces[0].PrivateIpAddresses[] | select(.Primary==false) | .PrivateIpAddress'`
+}
+
 #### Global variables - These allow the script to be run by non-bigbang devs easily
 if [[ -z "${VPC_ID}" ]]; then
   # default
@@ -84,7 +88,9 @@ while [ -n "$1" ]; do # while loop starts
       METAL_LB=true
   ;;
 
-  -a) echo "-a option passed to create secondary public IP"
+  -a) echo "-a option passed to create secondary public IP (-p and -m flags are skipped if set)"
+      PRIVATE_IP=false
+      METAL_LB=false
       ATTACH_SECONDARY_IP=true
   ;;
 
@@ -105,7 +111,7 @@ while [ -n "$1" ]; do # while loop starts
         fi
 
         aws ec2 terminate-instances --instance-ids ${AWSINSTANCEIDs} &>/dev/null
-        echo -n "waiting for instance termination..."
+        echo -n "Waiting for instance termination..."
         aws ec2 wait instance-terminated --instance-ids ${AWSINSTANCEIDs} &> /dev/null
         echo "done"
       else
@@ -131,7 +137,7 @@ while [ -n "$1" ]; do # while loop starts
       echo " -b   use BIG M5 instance. Default is t3.2xlarge"
       echo " -p   use private IP for security group and k3d cluster"
       echo " -m   create k3d cluster with metalLB"
-      echo " -a   attach secondary Public IP"
+      echo " -a   attach secondary Public IP (overrides -p and -m flags)"
       echo " -d   destroy related AWS resources"
       echo " -h   output help"
       exit 0
@@ -150,7 +156,8 @@ InstId=`aws ec2 describe-instances \
         --filters "Name=tag:Name,Values=${AWSUSERNAME}-dev" "Name=instance-state-name,Values=running"`
   if [[ ! -z "${InstId}" ]]; then
     PublicIP=`aws ec2 describe-instances --output text --no-cli-pager --instance-id ${InstId} --query "Reservations[].Instances[].PublicIpAddress"`
-    echo "Existing cluster found running on instance ${InstId} on ${PublicIP}"
+    PrivateIP=`aws ec2 describe-instances --output json --no-cli-pager --instance-ids ${InstId} | jq -r '.Reservations[0].Instances[0].PrivateIpAddress'`
+    echo "Existing cluster found running on instance ${InstId} on ${PublicIP} / ${PrivateIP}"
     echo "💣 Big Bang Cluster Management 💣"
     PS3="Please select an option: "
     options=("Re-create K3D cluster" "Recreate the EC2 instance from scratch" "Quit")
@@ -166,6 +173,12 @@ InstId=`aws ec2 describe-instances \
             exit 1
           fi
           RESET_K3D=true
+          SecondaryIP=`aws ec2 describe-instances --output json --no-cli-pager --instance-ids ${InstId} | jq -r '.Reservations[0].Instances[0].NetworkInterfaces[0].PrivateIpAddresses[] | select(.Primary==false) | .Association.PublicIp'`
+          PrivateIP2=$(getPrivateIP2)
+          if [[ "${ATTACH_SECONDARY_IP}" == true && -z "${SecondaryIP}" ]]; then
+            echo "Secondary IP didn't exist at the time of creation of the instance, so cannot attach one without re-creating it with the -a flag selected."
+            exit 1
+          fi
           run "k3d cluster delete"
           break;;
         2)
@@ -178,6 +191,11 @@ InstId=`aws ec2 describe-instances \
 
           aws ec2 terminate-instances --instance-ids ${InstId} &>/dev/null
           echo -n "Instance is being terminated..."
+          if [[ "${ATTACH_SECONDARY_IP}" == true ]]; then
+            echo -n "Waiting for instance termination..."
+            aws ec2 wait instance-terminated --instance-ids ${InstId} &> /dev/null
+            echo "done"
+          fi
           break;;
         3)
           echo "Bye."
@@ -253,7 +271,7 @@ if [[ "${RESET_K3D}" == false ]]; then
     echo -e "missing\nAdding ${WorkstationIP} to security group ${SGname} ..."
     if [[ "$PRIVATE_IP" == true ]];
     then
-        aws ec2 authorize-security-group-ingress --output json --no-cli-pager --group-name ${SGname} --protocol tcp --port 22 --cidr ${WorkstationIP}/32
+      aws ec2 authorize-security-group-ingress --output json --no-cli-pager --group-name ${SGname} --protocol tcp --port 22 --cidr ${WorkstationIP}/32
     else  # all protocols to all ports is the default
       aws ec2 authorize-security-group-ingress --output json --no-cli-pager --group-name ${SGname} --protocol all --cidr ${WorkstationIP}/32
     fi
@@ -385,9 +403,6 @@ EOF
   ## IP Address Allocation and Attachment
   CURRENT_EPOCH=`date +'%s'`
 
-  # Get the public IP address of our instance
-  PublicIP=`aws ec2 describe-instances --output json --no-cli-pager --instance-ids ${InstId} | jq -r '.Reservations[0].Instances[0].PublicIpAddress'`
-
   # Get the private IP address of our instance
   PrivateIP=`aws ec2 describe-instances --output json --no-cli-pager --instance-ids ${InstId} | jq -r '.Reservations[0].Instances[0].PrivateIpAddress'`
 
@@ -404,16 +419,13 @@ EOF
       echo "Previously allocated primary Elastic IP ${PublicIP} found."
     fi
 
-    echo "Instance public IP is ${PublicIP}"
-    echo
-
     echo -n "Associating IP ${PublicIP} address to instance ${InstId} ..."
     EIP1_ASSOCIATION_ID=`aws ec2 associate-address --output json --no-cli-pager --instance-id ${InstId} --private-ip ${PrivateIP} --public-ip $PublicIP | jq -r '.AssociationId'`
     echo "${EIP1_ASSOCIATION_ID}"
     EIP1_ID=`aws ec2 describe-addresses --public-ips ${PublicIP} | jq -r '.Addresses[].AllocationId'`
     aws ec2 create-tags --resources ${EIP1_ID} --tags Key="lastused",Value="${CURRENT_EPOCH}"
 
-    PrivateIP2=`aws ec2 describe-instances --output json --no-cli-pager --instance-ids ${InstId} | jq -r '.Reservations[0].Instances[0].NetworkInterfaces[0].PrivateIpAddresses[] | select(.Primary==false) | .PrivateIpAddress'`
+    PrivateIP2=$(getPrivateIP2)
     echo "Checking to see if a Secondary Elastic IP is already allocated and not attached..."
     SecondaryIP=`aws ec2 describe-addresses --filter "Name=tag:Name,Values=${AWSUSERNAME}-EIP2" --query 'Addresses[?AssociationId==null]' | jq -r '.[0].PublicIp // ""'`
     if [[ -z "${SecondaryIP}" ]]; then
@@ -429,6 +441,24 @@ EOF
     aws ec2 create-tags --resources ${EIP2_ID} --tags Key="lastused",Value="${CURRENT_EPOCH}"
     echo "Secondary public IP is ${SecondaryIP}"
   fi
+
+  echo
+  echo "Instance ${InstId} is ready!"
+  echo "Instance Public IP is ${PublicIP}"
+  echo "Instance Private IP is ${PrivateIP}"
+  echo
+
+  # Remove previous keys related to this IP from your SSH known hosts so you don't end up with a conflict
+  ssh-keygen -f "${HOME}/.ssh/known_hosts" -R "${PublicIP}"
+
+  echo "ssh init"
+  # this is a do-nothing remote ssh command just to initialize ssh and make sure that the connection is working
+  until run "hostname"; do
+    sleep 5
+    echo "Retry ssh command.."
+  done
+  echo
+
   ##### Configure Instance
   ## TODO: replace these individual commands with userdata when the spot instance is created?
   echo
@@ -465,26 +495,6 @@ EOF
   run 'sudo mv /home/ubuntu/kubectl /usr/local/bin/'
   run 'sudo chmod +x /usr/local/bin/kubectl'
 
-  # Get the private IP address of our instance
-  PrivateIP=`aws ec2 describe-instances --output json --no-cli-pager --instance-ids ${InstId} | jq -r '.Reservations[0].Instances[0].PrivateIpAddress'`
-
-  echo
-  echo "Instance ${InstId} is ready!"
-  echo "Instance private IP is ${PrivateIP}"
-  echo "Instance public IP is ${PublicIP}"
-  echo
-
-  # Remove previous keys related to this IP from your SSH known hosts so you don't end up with a conflict
-  ssh-keygen -f "${HOME}/.ssh/known_hosts" -R "${PublicIP}"
-
-  echo "ssh init"
-  # this is a do-nothing remote ssh command just to initialize ssh and make sure that the connection is working
-  until run "hostname"; do
-    sleep 5
-    echo "Retry ssh command.."
-  done
-  echo
-
   echo
   echo
   # install k3d on instance
@@ -505,18 +515,19 @@ k3d_command="k3d cluster create --servers 1 --agents 3"
 k3d_command+=" -v /etc:/etc@server:*\;agent:* -v /dev/log:/dev/log@server:*\;agent:* -v /run/systemd/private:/run/systemd/private@server:*\;agent:*"
 # Disable traefik and metrics-server
 k3d_command+=" --k3s-arg \"--disable=traefik@server:0\" --k3s-arg \"--disable=metrics-server@server:0\""
-# Port mappings to support Istio ingress + API access
 
+# Port mappings to support Istio ingress + API access
+if [[ -z "${PrivateIP2}" ]]; then
+  k3d_command+=" --port ${PrivateIP}:80:80@loadbalancer --port ${PrivateIP}:443:443@loadbalancer --api-port 6443"
+else
+  k3d_command+=" --port ${PrivateIP}:80:80@loadbalancer --port ${PrivateIP}:443:443@loadbalancer --port ${PrivateIP2}:80:81@loadbalancer --port ${PrivateIP2}:443:444@loadbalancer --api-port 6443"
+fi
+
+# Selecting K8S version through the use of a K3S image tag
 K3S_IMAGE_TAG=${K3S_IMAGE_TAG:=""}
 if [[ ! -z "$K3S_IMAGE_TAG" ]]; then
   echo "Using custom K3S image tag $K3S_IMAGE_TAG..."
   k3d_command+=" --image docker.io/rancher/k3s:$K3S_IMAGE_TAG"
-fi
-
-if [[ "${ATTACH_SECONDARY_IP}" == false ]]; then
-  k3d_command+=" --port ${PrivateIP}:80:80@loadbalancer --port ${PrivateIP}:443:443@loadbalancer --api-port 6443"
-else
-  k3d_command+=" --port ${PrivateIP}:80:80@loadbalancer --port ${PrivateIP}:443:443@loadbalancer --port ${PrivateIP2}:80:81@loadbalancer --port ${PrivateIP2}:443:444@loadbalancer --api-port 6443"
 fi
 
 # Add MetalLB specific k3d config
@@ -668,7 +679,7 @@ else   # Not using MetalLB and using pubilc IP. This is the default
   echo
 fi
 
-if [[ "${ATTACH_SECONDARY_IP}" == true ]]; then
+if [[ ! -z "${SecondaryIP}" ]]; then
   echo "A secondary IP is available for use if you wish to have a passthrough ingress for Istio along with a public Ingress Gateway, this maybe useful for Keycloak x509 mTLS authentication."
   echo "  ${SecondaryIP}  keycloak.bigbang.dev"
 fi
